src/third_party/blink/web_tests/platform/win/external/wpt/mathml/presentation-markup/scripts/subsup-5-expected.txt

This is a testharness.js-based test.
PASS Alignment on the baseline with different and large script heights
FAIL Tall subscripts/superscripts are not placed too high/low assert_less_than: mmultiscripts: postsupscript is above the bottom of the base expected a number less than 184.1875 but got 434.1875
FAIL No collisions for tall subscripts and superscripts assert_greater_than: mmultiscripts: presubscript is below the presuperscript expected a number greater than 684.1875 but got 434.1875
Harness: the test ran to completion.
